@@ -31,21 +31,24 @@ line_length = 79
 
 [tool.poetry]
 name = "abses"
-version = "0.1.3"
+version = "0.2.0.alpha"
 description = "ABSESpy makes it easier to build artificial Social-ecological systems with real GeoSpatial datasets and fully incorporate human behaviour."
 authors = ["Shuang Song <songshgeo@gmail.com>"]
 license = "Apache 2.0 License"
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = ">3.8.1"
+python = ">=3.9,<4"
 prettytable = "^3.6.0"
 pint = "^0.20.1"
 netcdf4 = "^1.6.2"
-agentpy = "^0.1.5"
+hydra-core = "^1.3.2"
+mesa-geo = "^0.5.0"
+xarray = "^2023.8.0"
+fiona = "^1.9.4.post1"
 
 [tool.poetry.group.dev.dependencies]
-agentpy = "^0.1.5"
+pytest-clarity = "^1.0.1"
 numpy = "^1.24.1"
 pandas = "^1.5.3"
 xarray = "^2023.1.0"
@@ -55,7 +58,44 @@ pre-commit = "^3.0.1"
 scriv = "^1.2.0"
 pytest = "^7.2.1"
 sourcery = "^1.0.6"
+allure-pytest = "^2.13.2"
+pytest-sugar = "^0.9.7"
+ipykernel = "^6.25.1"
+jupyterlab = "^4.0.5"
+jupyterlab-execute-time = "^3.0.1"
+matplotlib = "^3.7.2"
+pytest-cov = "^4.1.0"
+flake8 = "^6.1.0"
+isort = "^5.12.0"
+nbstripout = "^0.6.1"
+pydocstyle = "^6.3.0"
+pre-commit-hooks = "^4.4.0"
+interrogate = "^1.5.0"
+mypy = "^1.5.1"
+bandit = "^1.7.5"
+black = "^23.9.1"
+pylint = "^3.0.1"
+
+
+[tool.poetry.group.docs.dependencies]
+mkdocs = "^1.5.3"
+mkdocs-material = "^9.4.2"
+mkdocs-git-revision-date-localized-plugin = "^1.2.0"
+mkdocs-minify-plugin = "^0.7.1"
+mkdocs-redirects = "^1.2.1"
+mkdocs-awesome-pages-plugin = "^2.9.2"
+mkdocs-git-authors-plugin = "^0.7.2"
+mkdocstrings = {extras = ["python"], version = "^0.23.0"}
+mkdocs-bibtex = "^2.11.0"
+mkdocs-macros-plugin = "^1.0.4"
+mkdocs-jupyter = "^0.24.5"
+mkdocs-callouts = "^1.9.1"
+mkdocs-glightbox = "^0.3.4"
 
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.scriv]
+version = "literal: pyproject.toml: tool.poetry.version"
+format = "md"
