@@ -5,172 +5,70 @@
 # GitHub   : https://github.com/SongshGeo
 # Website: https://cv.songshgeo.com/
 
-from abc import abstractmethod
-from typing import Iterable, List, Optional, Union
+from __future__ import annotations
 
-from agentpy.tools import AttrDict
+from typing import TYPE_CHECKING, Iterable, Optional, Set, Union
 
-from .bases import Mediator
-from .log import Log
-from .tools.func import iter_func, make_list
+from omegaconf import DictConfig
 
-STATES = {
-    -1: "Waiting",
-    0: "new",
-    1: "init",
-    2: "ready",
-    3: "complete",
-}
+from .log import _Log
+from .tools.func import make_list
+
+if TYPE_CHECKING:
+    from abses.main import MainModel
 
 
-class Component(Log):
+class _Component(_Log):
     """
-    The Base Component provides the basic functionality of storing a mediator's
-    instance inside component objects.
+    Class _Component is used to represent a component of the model. Extends _Log class to be able to log any issues.
+    Components are foundational pieces in constructing the model's entities.
+    It is initialized with a model and a optional name.
     """
 
-    def __init__(self, name: Optional[str] = None):
-        Log.__init__(self, name=name)
-        self._arguments: List[str] = []
-        self._parameters: dict = AttrDict()
+    __args__ = []
+
+    def __init__(self, model: MainModel, name: Optional[str] = None):
+        _Log.__init__(self, name=name)
+        self._args: Set[str] = set()
+        self._model = model
+        self.add_args(self.__args__)
 
     @property
     def params(self) -> dict:
-        return self._parameters
-
-    @params.setter
-    def params(self, my_params: dict) -> None:
-        self.params.update(my_params)
-
-    @property
-    def arguments(self) -> List[str]:
-        return self._arguments
-
-    @arguments.setter
-    def arguments(self, args: "str|Iterable[str]") -> None:
-        arg_lst = make_list(args)
-        self._arguments.extend(arg_lst)
-        self._arguments = sorted(list(set(self.arguments)))
-
-    @iter_func("modules")
-    def _parsing_args(self) -> None:
-        """
-        Initialize arguments of the component then delete from param dictionary.
-
-        Raises:
-            KeyError: cannot find argument in parameters.
-        """
-        arguments = getattr(self.__class__, "args", [])
-        self.arguments = arguments
-        for arg in self.arguments:
-            try:
-                value = self.params[arg]
-                self.__setattr__(arg, value)
-                del self.params[arg]
-            except KeyError as e:
-                raise KeyError(f"arg '{arg}' not found in parameters.") from e
-
-    @iter_func("modules")
-    def parsing_params(self, params: dict) -> dict:
-        """
-        Parsing parameters belongs to this component.
-
-        Args:
-            params (dict): input settings.
+        """Returns read-only model's parameters.
 
         Returns:
-            dict: unsolved parameters.
+        ___________
+        dict:
+            Dictionary of model's parameters.
         """
-        # select my params
-        self.params = params.pop(self.name, {})
-        # retrieve specific parameters and update
-        for key in list(params.keys()):
-            if key in self.params:
-                value = params.pop(key)
-                self.logger.debug(
-                    "Using %s to update %s for '%s'",
-                    value,
-                    self.params[key],
-                    key,
-                )
-                self.params[key] = value
-        # setup arguments
-        self._parsing_args()
-        # solving parameters diversely
-        self._after_parsing()
-        # handle parameters defined by users
-        self.handle_params()
-        return params
-
-    def report_parameters(
-        self, params: Optional[dict] = None, as_string: Optional[bool] = False
-    ) -> Union[dict, str]:
-        # TODO: show parameter table.
-        if len(params) == 0:
-            return None
-        return str(list(params.keys())) if as_string else params.keys()
-
-    @abstractmethod
-    def _after_parsing(self):
-        pass
-
-    @abstractmethod
-    def handle_params(self):
-        """
-        Handle parameters after loading.
-        """
-
-    @abstractmethod
-    def initialize(self):
-        """
-        Initialization after handle parameters.
-        """
-
-
-class MainComponent(Component):
-    _states = STATES
-
-    def __init__(self, name: Optional[str] = None):
-        Component.__init__(self, name=name)
-        self._state = -1  # init state waiting
-        self._mediator = Mediator()
+        return self._model.settings.get(self.name, DictConfig({}))
 
     @property
-    def mediator(self) -> Mediator:
-        return self._mediator
+    def args(self) -> DictConfig:
+        """Returns read-only component's arguments.
 
-    @mediator.setter
-    def mediator(self, mediator) -> None:
-        self._mediator = mediator
+        Returns
+        -------
+        DictConfig
+            Component's arguments dictionary.
+        """
+        return DictConfig({arg: self.params[arg] for arg in self._args})
 
-    @property
-    def state(self) -> str:
-        return self._states[self._state]
+    def add_args(self, args: Union[str, Iterable[str]]) -> None:
+        """Add model's parameters as component's arguments.
 
-    @state.setter
-    def state(self, code):
-        if code not in self._states:
-            raise ValueError(f"Invalid state {code}, valid: {self._states}!")
-        elif code == self._state:
-            raise ValueError(f"Setting state repeat: {self.state}!")
-        elif code < self._state:
-            raise ValueError(
-                f"State cannot retreat from {self._states[code]} to {self.state}!"
-            )
-        else:
-            self._state = code
-        self.mediator.transfer_event(sender=self, event=self.state)
+        Parameters
+        ----------
+        args : Union[str, Iterable[str]]
+            Model's parameters to be added as component's arguments.
 
-    def parsing_params(self, params: dict) -> dict:
-        unsolved = super().parsing_params(params)
-        if self.mediator.transfer_parsing(self, params):
-            parameters = self.report_parameters(unsolved, as_string=True)
-            message = f"Unsolved parameters: {parameters}."
-            self.mediator.logging(
-                message, condition=len(unsolved) > 0, level="warning"
-            )
-        return unsolved
-
-    @iter_func("modules")
-    def _after_parsing(self):
-        pass
+        Returns
+        _______
+        None
+        """
+        args_set = set(make_list(args))
+        for arg in args_set:
+            if arg not in self.params:
+                raise KeyError(f"Argument {arg} not found.")
+            self._args.add(arg)
