@@ -11,8 +11,12 @@ import logging.config
 from .tools.func import iter_func
 
 
+# https://towardsdatascience.com/5-easy-and-effective-ways-to-use-python-logging-a9564bd17ccd
+# TODO 这里可以使用更好的工作流进行日志记录
 # TODO change a specific log for experiments
-class Log(object):
+class _Log(object):
+    """日志记录器"""
+
     def __init__(self, name):
         if name is None:
             name = self.__class__.__name__.lower()
