@@ -8,8 +8,31 @@ import logging
 from typing import Any, Iterable, List
 
 import numpy as np
+from scipy import ndimage
 
 logger = logging.getLogger(__name__)
+
+
+def get_buffer(
+    array: np.ndarray,
+    radius: int = 1,
+    moor: bool = False,
+    annular: bool = False,
+) -> np.ndarray:
+    """Get a buffer around the array."""
+    if radius <= 0:
+        raise ValueError(f"Radius must be positive, not {radius}.")
+    connectivity = 2 if moor else 1
+    struct = ndimage.generate_binary_structure(2, connectivity)
+    result = ndimage.binary_dilation(
+        array, structure=struct, iterations=radius
+    )
+    if annular and radius > 1:
+        interior = ndimage.binary_dilation(
+            array, structure=struct, iterations=radius - 1
+        )
+        return result & ~interior
+    return result
 
 
 def make_list(element, keep_none=False):
@@ -94,130 +117,3 @@ def iter_func(elements: str) -> callable:
         return broadcast_func
 
     return broadcast
-
-
-def opfunc_using_attr(attr: str, binary_op: bool) -> callable:
-    """
-    wrap a operation method of a custom class, for using its 'attribution' in calculation.
-
-    Args:
-        attr (str): which attribute of the class to use in calculation.
-        binary_op (bool): if the wrapping operation method is binary or unary.
-
-    Returns:
-        callable: a wrapped bound method.
-
-    Examples:
-        class Calculation:
-            def __init__(self):
-                self.attr = 1
-
-        cal = Calculation()
-
-        cal + 1
-        >>> TypeError: 'unsupported operand type(s)' ...
-
-        # wraping the '__add__' method.
-        opname = '__add__'
-        wrapper = opfunc_using_attr('attr', binary_op=True)
-        wrapped_func = wrapper(getattr(1, opname))
-        setattr(cal.__class__, opname, wrapped_func)
-
-        cal + 1 == 2
-        >>> 2
-    """
-
-    # wraping binary operation functions.
-    def wrap_binary_opfunc(func):
-        def new_func(self, other):
-            opfunc = getattr(getattr(self, attr), func.__name__)
-            if issubclass(other.__class__, self.__class__):
-                return opfunc(getattr(other, attr))
-            else:
-                return opfunc(other)
-
-        return new_func
-
-    # wraping unary operation functions.
-    def wrap_unary_opfunc(func):
-        def new_func(self):
-            opfunc = getattr(getattr(self, attr), func.__name__)
-            return opfunc()
-
-        return new_func
-
-    return wrap_binary_opfunc if binary_op else wrap_unary_opfunc
-
-
-# https://zhuanlan.zhihu.com/p/515284250
-_binary_opfunc = (
-    "__add__",
-    "__sub__",
-    "__mul__",
-    "__truediv__",
-    "__floordiv__",
-    "__mod__",
-    "__divmod__",
-    "__pow__",
-    "__radd__",
-    "__rsub__",
-    "__rmul__",
-    "__rtruediv__",
-    "__rfloordiv__",
-    "__rmod__",
-    "__rdivmod__",
-    "__rpow__",
-    "__lshift__",
-    "__rshift__",
-    "__and__",
-    "__xor__",
-    "__or__",
-    "__contains__",
-)
-
-_unary_opfunc = (
-    "__neg__",
-    "__pos__",
-    "__abs__",
-    "__invert__",
-)
-
-
-def wrap_opfunc_to(obj: object, attr: str) -> List[str]:
-    """
-    Overwrite an object's all operation functions, using its attribute in future's calculations.
-
-    Args:
-        obj (object): any object to be wrapped.
-        attr (str): object's attribute name.
-
-    Returns:
-        List[str]: successful wrapped operation functions' name.
-
-    Examples:
-        class MyClass(object):
-            def __init__(self, attr: any):
-                self.attr: any = attr
-        int_obj = MyClass(1)
-        str_obj = MyClass('test')
-        wrap_opfunc_to
-    """
-
-    def wrap_opfunc(binary_op):
-        func = getattr(getattr(obj, attr), opname, None)
-        if func is None:
-            return False
-        wrapper = opfunc_using_attr(attr, binary_op=binary_op)
-        setattr(obj.__class__, opname, wrapper(func))
-        return True
-
-    successful_wrapped = []
-    for opname in _binary_opfunc:
-        wrapped = wrap_opfunc(True)
-        if wrapped:
-            successful_wrapped.append(opname.strip("_"))
-    for opname in _unary_opfunc:
-        wrapped = wrap_opfunc(False)
-        if wrapped:
-            successful_wrapped.append(opname.strip("_"))
-    return successful_wrapped
