@@ -7,9 +7,10 @@
 
 import numpy as np
 
-from abses.human import HumanModule
+from abses import MainModel
+from abses.nature import PatchModule
 
-from .create_tested_instances import Admin, Farmer, simple_main_model
+from .fixtures import Admin, Farmer
 
 selection = {
     "breed": "Farmer",
@@ -18,18 +19,17 @@ selection = {
 
 
 def test_human_attributes():
-    model = simple_main_model()
+    model = MainModel()
     human = model.human
     assert human.agents is model.agents
-    assert len(human.actors) == 0
+    assert len(human.actors()) == 0
     human.agents.create(Farmer, 5)
     model.agents.create(Admin, 5)
     assert len(human.agents) == 10
-    assert len(human.actors.now()) == 0
 
 
 def test_human_define():
-    model = simple_main_model()
+    model = MainModel()
     human = model.human
     farmers = human.agents.create(Farmer, 5)
     admins = model.agents.create(Admin, 5)
@@ -38,24 +38,25 @@ def test_human_define():
     admins.update("test", np.arange(5))
 
     human.define("test", "Farmer")
-    assert human.test == farmers
+    assert human.actors("test") == farmers
     human.agents.remove(farmers[0])
-    assert human.test == farmers[1:]
+    assert human.actors("test") == farmers[1:]
 
 
 def test_human_rule():
-    model = simple_main_model(test_human_rule)
+    model = MainModel()
+    layer = PatchModule.from_resolution(model)
     human = model.human
     farmers = human.agents.create(Farmer, 5)
-    farmers.rule(
-        when="test == 1", then="settle_down", position=(3, 3), disposable=True
+    farmers.trigger(
+        "put_on_layer",
+        layer=layer,
+        pos=(2, 3),
     )
+
+    farmers.trigger("rule", when="test == 1", then="die", disposable=True)
     # assert checked
     farmers.update("test", np.arange(5))
-    assert human.agents.to_list().on_earth.sum() == 1
-    farmers.rule(when="test == 2", then="die")
-    assert human.agents.__len__() == 4
-
-
-def test_human_arena():
-    pass
+    assert human.agents.to_list().on_earth.sum() == 4
+    farmers.trigger("rule", when="test == 2", then="die")
+    assert len(human.agents) == 3
